refactor(step2): 패키지 이동
