+++ src/main/java/nextstep/subway/config/ReplicationRoutingDataSource.java
@@ -1,4 +1,4 @@
-package nextstep.subway;
+package nextstep.subway.config;
 
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
